requirements.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile pyproject.toml
aiohappyeyeballs==2.6.1
    # via aiohttp
aiohttp==3.11.18
    # via fsspec
aiosignal==1.3.2
    # via aiohttp
antlr4-python3-runtime==4.9.3
    # via omegaconf
attrs==25.3.0
    # via aiohttp
backports-functools-lru-cache==2.0.0
    # via cpflows
bokeh==3.7.3
    # via dask
click==8.2.1
    # via
    #   dask
    #   distributed
cloudpickle==3.1.1
    # via
    #   dask
    #   distributed
contourpy==1.3.2
    # via
    #   bokeh
    #   matplotlib
cpflows @ git+https://github.com/Vekteur/CP-Flow.git@4aedb04d29ae4c0c24f211c184bf43ddd946e57f
    # via moc (pyproject.toml)
cycler==0.12.1
    # via
    #   cpflows
    #   matplotlib
dask==2025.5.1
    # via
    #   moc (pyproject.toml)
    #   distributed
distributed==2025.5.1
    # via dask
filelock==3.18.0
    # via torch
fonttools==4.58.0
    # via matplotlib
frozenlist==1.6.0
    # via
    #   aiohttp
    #   aiosignal
fsspec==2025.5.0
    # via
    #   dask
    #   lightning
    #   pytorch-lightning
    #   torch
future==1.0.0
    # via cpflows
h5py==3.13.0
    # via cpflows
idna==3.10
    # via yarl
jinja2==3.1.6
    # via
    #   bokeh
    #   dask
    #   distributed
    #   torch
joblib==1.5.0
    # via scikit-learn
kiwisolver==1.4.8
    # via
    #   cpflows
    #   matplotlib
lightning==2.5.1.post0
    # via moc (pyproject.toml)
lightning-utilities==0.14.3
    # via
    #   lightning
    #   pytorch-lightning
    #   torchmetrics
locket==1.0.0
    # via
    #   distributed
    #   partd
lz4==4.4.4
    # via dask
markdown-it-py==3.0.0
    # via rich
markupsafe==3.0.2
    # via jinja2
matplotlib==3.10.3
    # via
    #   moc (pyproject.toml)
    #   cpflows
    #   seaborn
mdurl==0.1.2
    # via markdown-it-py
mpmath==1.3.0
    # via sympy
msgpack==1.1.0
    # via distributed
multidict==6.4.4
    # via
    #   aiohttp
    #   yarl
narwhals==1.40.0
    # via bokeh
networkx==3.4.2
    # via torch
numpy==2.2.6
    # via
    #   moc (pyproject.toml)
    #   bokeh
    #   contourpy
    #   cpflows
    #   dask
    #   h5py
    #   matplotlib
    #   pandas
    #   pyro-ppl
    #   scikit-learn
    #   scipy
    #   seaborn
    #   torchmetrics
    #   torchvision
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
omegaconf==2.3.0
    # via moc (pyproject.toml)
opt-einsum==3.4.0
    # via pyro-ppl
packaging==24.2
    # via
    #   bokeh
    #   dask
    #   distributed
    #   lightning
    #   lightning-utilities
    #   matplotlib
    #   pytorch-lightning
    #   torchmetrics
pandas==2.2.3
    # via
    #   bokeh
    #   cpflows
    #   dask
    #   seaborn
partd==1.4.2
    # via dask
pillow==11.2.1
    # via
    #   bokeh
    #   matplotlib
    #   torchvision
propcache==0.3.1
    # via
    #   aiohttp
    #   yarl
psutil==7.0.0
    # via distributed
pyarrow==20.0.0
    # via dask
pygments==2.19.1
    # via rich
pyparsing==3.2.3
    # via
    #   cpflows
    #   matplotlib
pyro-api==0.1.2
    # via pyro-ppl
pyro-ppl==1.9.1
    # via moc (pyproject.toml)
python-dateutil==2.9.0.post0
    # via
    #   cpflows
    #   matplotlib
    #   pandas
pytorch-lightning==2.5.1.post0
    # via lightning
pytz==2025.2
    # via
    #   cpflows
    #   pandas
pyyaml==6.0.2
    # via
    #   moc (pyproject.toml)
    #   bokeh
    #   dask
    #   distributed
    #   lightning
    #   omegaconf
    #   pytorch-lightning
rich==14.0.0
    # via moc (pyproject.toml)
scikit-learn==1.6.1
    # via
    #   moc (pyproject.toml)
    #   cpflows
scipy==1.15.3
    # via
    #   moc (pyproject.toml)
    #   cpflows
    #   scikit-learn
seaborn==0.13.2
    # via
    #   moc (pyproject.toml)
    #   cpflows
setuptools==80.8.0
    # via
    #   lightning-utilities
    #   torch
    #   triton
six==1.17.0
    # via
    #   cpflows
    #   python-dateutil
sortedcontainers==2.4.0
    # via distributed
subprocess32==3.5.4
    # via cpflows
sympy==1.14.0
    # via torch
tblib==3.1.0
    # via distributed
threadpoolctl==3.6.0
    # via scikit-learn
toolz==1.0.0
    # via
    #   dask
    #   distributed
    #   partd
torch==2.7.0
    # via
    #   moc (pyproject.toml)
    #   cpflows
    #   lightning
    #   pyro-ppl
    #   pytorch-lightning
    #   torchmetrics
    #   torchvision
torchmetrics==1.7.1
    # via
    #   lightning
    #   pytorch-lightning
torchvision==0.22.0
    # via cpflows
tornado==6.5
    # via
    #   bokeh
    #   distributed
tqdm==4.67.1
    # via
    #   moc (pyproject.toml)
    #   cpflows
    #   lightning
    #   pyro-ppl
    #   pytorch-lightning
triton==3.3.0
    # via torch
typing-extensions==4.13.2
    # via
    #   lightning
    #   lightning-utilities
    #   pytorch-lightning
    #   torch
tzdata==2025.2
    # via pandas
urllib3==2.4.0
    # via distributed
xyzservices==2025.4.0
    # via bokeh
yarl==1.20.0
    # via aiohttp
zict==3.0.0
    # via distributed
